EXP-452 EXP-939 refactor channel, rename model abstract class
