+++ app/src/main/java/com/scala/expandroidsdk/model/Data.java
@@ -3,6 +3,6 @@ package com.scala.expandroidsdk.model;
 /**
  * Created by Cesar Oyarzun on 10/28/15.
  */
-public class Data extends AbsModel{
+public class Data extends AbstractModel {
 
 }
