+++ app/src/main/java/com/scala/expandroidsdk/model/Device.java
@@ -3,6 +3,6 @@ package com.scala.expandroidsdk.model;
 /**
  * Created by Cesar Oyarzun on 10/30/15.
  */
-public class Device extends AbsModel {
+public class Device extends AbstractModel {
 
 }
