+++ app/src/main/java/com/scala/expandroidsdk/model/Experience.java
@@ -3,6 +3,6 @@ package com.scala.expandroidsdk.model;
 /**
  * Created by Cesar Oyarzun on 10/30/15.
  */
-public class Experience extends AbsModel{
+public class Experience extends AbstractModel {
 
 }
