+++ app/src/main/java/com/scala/expandroidsdk/model/Location.java
@@ -4,6 +4,6 @@ package com.scala.expandroidsdk.model;
  * Created by Cesar Oyarzun on 10/28/15.
  */
 
-public class Location extends AbsModel {
+public class Location extends AbstractModel {
 
 }
