+++ app/src/main/java/com/scala/expandroidsdk/model/Thing.java
@@ -3,6 +3,6 @@ package com.scala.expandroidsdk.model;
 /**
  * Created by Cesar Oyarzun on 10/28/15.
  */
-public class Thing extends AbsModel{
+public class Thing extends AbstractModel {
 
 }
